requirements.txt
openai-whisper
watchdog
python-dotenv